Revert "Break up ChainBuilder logic"
This reverts commit 54230c305655690043978c82399255c3377b18c3.
