+++ ethereum/core/src/testFixtures/java/tech/pegasys/teku/core/ChainBuilder.java
@@ -19,8 +19,14 @@ import static tech.pegasys.teku.datastructures.util.BeaconStateUtil.compute_star
 import static tech.pegasys.teku.util.config.Constants.SLOTS_PER_EPOCH;
 
 import com.google.common.primitives.UnsignedLong;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.NavigableMap;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.TreeMap;
 import java.util.stream.LongStream;
 import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes32;
@@ -35,6 +41,7 @@ import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.operations.Attestation;
 import tech.pegasys.teku.datastructures.operations.DepositData;
 import tech.pegasys.teku.datastructures.state.BeaconState;
+import tech.pegasys.teku.datastructures.state.Checkpoint;
 import tech.pegasys.teku.datastructures.util.DepositGenerator;
 import tech.pegasys.teku.datastructures.util.MockStartBeaconStateGenerator;
 import tech.pegasys.teku.datastructures.util.MockStartDepositGenerator;
@@ -42,24 +49,26 @@ import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZMutableList;
 import tech.pegasys.teku.util.config.Constants;
 
-/** A utility for generating small, valid chains of blocks with states for testing */
-public class ChainBuilder extends DelegatingEth2Chain {
+/** A utility for building small, valid chains of blocks with states for testing */
+public class ChainBuilder {
 
   private final List<BLSKeyPair> validatorKeys;
   private final AttestationGenerator attestationGenerator;
+  private final NavigableMap<UnsignedLong, SignedBlockAndState> blocks = new TreeMap<>();
 
   private BlockProposalTestUtil blockProposalTestUtil = new BlockProposalTestUtil();
 
-  private ChainBuilder(final List<BLSKeyPair> validatorKeys, final Eth2Chain existingChain) {
-    super(new SimpleEth2Chain());
+  private ChainBuilder(
+      final List<BLSKeyPair> validatorKeys,
+      final Map<UnsignedLong, SignedBlockAndState> existingBlocks) {
     this.validatorKeys = validatorKeys;
 
     attestationGenerator = new AttestationGenerator(validatorKeys);
-    eth2Chain.putAll(existingChain);
+    blocks.putAll(existingBlocks);
   }
 
   public static ChainBuilder create(final List<BLSKeyPair> validatorKeys) {
-    return new ChainBuilder(validatorKeys, new SimpleEth2Chain());
+    return new ChainBuilder(validatorKeys, Collections.emptyMap());
   }
 
   /**
@@ -69,11 +78,105 @@ public class ChainBuilder extends DelegatingEth2Chain {
    * @return An independent copy of this ChainBuilder
    */
   public ChainBuilder fork() {
-    return new ChainBuilder(validatorKeys, eth2Chain);
+    return new ChainBuilder(validatorKeys, blocks);
+  }
+
+  public UnsignedLong getLatestSlot() {
+    asserChainIsNotEmpty();
+    return getLatestBlockAndState().getBlock().getSlot();
+  }
+
+  public UnsignedLong getLatestEpoch() {
+    asserChainIsNotEmpty();
+    final UnsignedLong slot = getLatestSlot();
+    return compute_epoch_at_slot(slot);
+  }
+
+  public Stream<SignedBlockAndState> streamBlocksAndStates() {
+    return blocks.values().stream();
+  }
+
+  public Stream<SignedBlockAndState> streamBlocksAndStates(final long fromSlot, final long toSlot) {
+    return streamBlocksAndStates(UnsignedLong.valueOf(fromSlot), UnsignedLong.valueOf(toSlot));
+  }
+
+  public Stream<SignedBlockAndState> streamBlocksAndStates(
+      final UnsignedLong fromSlot, final UnsignedLong toSlot) {
+    return blocks.values().stream()
+        .filter(b -> b.getBlock().getSlot().compareTo(fromSlot) >= 0)
+        .filter(b -> b.getBlock().getSlot().compareTo(toSlot) <= 0);
+  }
+
+  public Stream<SignedBlockAndState> streamBlocksAndStatesUpTo(final long toSlot) {
+    return streamBlocksAndStatesUpTo(UnsignedLong.valueOf(toSlot));
+  }
+
+  public Stream<SignedBlockAndState> streamBlocksAndStatesUpTo(final UnsignedLong toSlot) {
+    return blocks.values().stream().filter(b -> b.getBlock().getSlot().compareTo(toSlot) <= 0);
+  }
+
+  public SignedBlockAndState getGenesis() {
+    return Optional.ofNullable(blocks.firstEntry()).map(Entry::getValue).orElse(null);
+  }
+
+  public SignedBlockAndState getLatestBlockAndState() {
+    return Optional.ofNullable(blocks.lastEntry()).map(Entry::getValue).orElse(null);
+  }
+
+  public SignedBlockAndState getBlockAndStateAtSlot(final long slot) {
+    return getBlockAndStateAtSlot(UnsignedLong.valueOf(slot));
+  }
+
+  public SignedBlockAndState getBlockAndStateAtSlot(final UnsignedLong slot) {
+    return Optional.ofNullable(blocks.get(slot)).orElse(null);
+  }
+
+  public SignedBeaconBlock getBlockAtSlot(final long slot) {
+    return getBlockAtSlot(UnsignedLong.valueOf(slot));
+  }
+
+  public SignedBeaconBlock getBlockAtSlot(final UnsignedLong slot) {
+    return resultToBlock(getBlockAndStateAtSlot(slot));
+  }
+
+  public BeaconState getStateAtSlot(final long slot) {
+    return getStateAtSlot(UnsignedLong.valueOf(slot));
+  }
+
+  public BeaconState getStateAtSlot(final UnsignedLong slot) {
+    return resultToState(getBlockAndStateAtSlot(slot));
+  }
+
+  public SignedBlockAndState getLatestBlockAndStateAtSlot(final long slot) {
+    return getLatestBlockAndStateAtSlot(UnsignedLong.valueOf(slot));
+  }
+
+  public SignedBlockAndState getLatestBlockAndStateAtSlot(final UnsignedLong slot) {
+    return Optional.ofNullable(blocks.floorEntry(slot)).map(Entry::getValue).orElse(null);
+  }
+
+  public SignedBlockAndState getLatestBlockAndStateAtEpochBoundary(final long epoch) {
+    return getLatestBlockAndStateAtEpochBoundary(UnsignedLong.valueOf(epoch));
+  }
+
+  public SignedBlockAndState getLatestBlockAndStateAtEpochBoundary(final UnsignedLong epoch) {
+    asserChainIsNotEmpty();
+    final UnsignedLong slot = compute_start_slot_at_epoch(epoch);
+    return getLatestBlockAndStateAtSlot(slot);
+  }
+
+  public Checkpoint getCurrentCheckpointForEpoch(final long epoch) {
+    return getCurrentCheckpointForEpoch(UnsignedLong.valueOf(epoch));
+  }
+
+  public Checkpoint getCurrentCheckpointForEpoch(final UnsignedLong epoch) {
+    asserChainIsNotEmpty();
+    final SignedBeaconBlock block = getLatestBlockAndStateAtEpochBoundary(epoch).getBlock();
+    return new Checkpoint(epoch, block.getMessage().hash_tree_root());
   }
 
   public SignedBlockAndState generateGenesis() {
-    checkState(eth2Chain.isEmpty(), "Genesis already created");
+    checkState(blocks.isEmpty(), "Genesis already created");
     final UnsignedLong genesisSlot = UnsignedLong.valueOf(Constants.GENESIS_SLOT);
 
     // Generate genesis state
@@ -81,14 +184,14 @@ public class ChainBuilder extends DelegatingEth2Chain {
         new MockStartDepositGenerator(new DepositGenerator(true)).createDeposits(validatorKeys);
     final BeaconState genesisState =
         new MockStartBeaconStateGenerator()
-            .createInitialBeaconState(genesisSlot, initialDepositData);
+            .createInitialBeaconState(UnsignedLong.ZERO, initialDepositData);
 
     // Generate genesis block
     BeaconBlock genesisBlock = new BeaconBlock(genesisState.hash_tree_root());
     final SignedBeaconBlock signedBlock = new SignedBeaconBlock(genesisBlock, BLSSignature.empty());
 
     final SignedBlockAndState blockAndState = new SignedBlockAndState(signedBlock, genesisState);
-    eth2Chain.putBlockAndState(blockAndState);
+    blocks.put(genesisSlot, blockAndState);
     return blockAndState;
   }
 
@@ -175,8 +278,12 @@ public class ChainBuilder extends DelegatingEth2Chain {
         .flatMap(head -> attestationGenerator.streamAttestations(head, head.getSlot()));
   }
 
+  private void asserChainIsNotEmpty() {
+    checkState(!blocks.isEmpty(), "Unable to execute operation on empty chain");
+  }
+
   private void assertBlockCanBeGenerated() {
-    checkState(!eth2Chain.isEmpty(), "Genesis block must be created before blocks can be added.");
+    checkState(!blocks.isEmpty(), "Genesis block must be created before blocks can be added.");
   }
 
   private SignedBlockAndState appendNewBlockToChain(
@@ -191,8 +298,16 @@ public class ChainBuilder extends DelegatingEth2Chain {
         blockProposalTestUtil.createBlockWithAttestations(
             signer, slot, preState, parentRoot, options.getAttestations());
 
-    eth2Chain.putBlockAndState(nextBlockAndState);
+    blocks.put(slot, nextBlockAndState);
     return nextBlockAndState;
+  }
+
+  private BeaconState resultToState(final SignedBlockAndState result) {
+    return Optional.ofNullable(result).map(SignedBlockAndState::getState).orElse(null);
+  }
+
+  private SignedBeaconBlock resultToBlock(final SignedBlockAndState result) {
+    return Optional.ofNullable(result).map(SignedBlockAndState::getBlock).orElse(null);
   }
 
   private MessageSignerService getSigner(final int proposerIndex) {
