+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/blocks/SignedBlockAndState.java
@@ -41,10 +41,6 @@ public class SignedBlockAndState {
     return block.getRoot();
   }
 
-  public Bytes32 getParentRoot() {
-    return block.getParent_root();
-  }
-
   public UnsignedLong getSlot() {
     return getBlock().getSlot();
   }
